Gwt module modified and reorganized.
