+++ gwt/src/main/java/com/google/code/morphia/Morphia.gwt.xml
@@ -1,9 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 1.7.0//EN" "http://google-web-toolkit.googlecode.com/svn/tags/1.7.0/distro-source/core/src/gwt-module.dtd">
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 1.7.0//EN"
+        "http://google-web-toolkit.googlecode.com/svn/tags/1.7.0/distro-source/core/src/gwt-module.dtd">
 <module>
-<!--	<inherits name="com.google.gwt.user.User"/>-->
-	<super-source path="super/"/>
-	<source path="">
-<!--		<include name="**/*.java"/>-->
-	</source>
+    <super-source path="super/" />
+    <source path="" />
 </module>
