+++ gwt/src/main/super/com/google/code/morphia/super/com/google/code/morphia/Key_CustomFieldSerializer.java
@@ -2,6 +2,7 @@ package com.google.code.morphia;
 
 import java.io.Serializable;
 
+import com.google.code.morphia.Key;
 import com.google.gwt.user.client.rpc.SerializationException;
 import com.google.gwt.user.client.rpc.SerializationStreamReader;
 import com.google.gwt.user.client.rpc.SerializationStreamWriter;
